Break up dialog components
